SFR-672 Add docker configuration files
This adds basic configuration of a docker deployment setup for the storybook instances. This will eventually be used to create multiple development environments that can be viewed for testing purposes.

At the moment this is just comprised of a `Dockerfile` that can be used to create either the react or twig storybook instances and a `docker-compose.yml` file that stands up containers for both instances on the development branch.

The work that remains to be done with this is:
- Create a process for creating instances on any branch
- Set up a cluster on ECS that can host multiple versions of storybook
- Integrate the deployment with travisCI