SFR-672 Add scripts for creating dev environments
To facilitate review by the design team and other interested stakeholders this adds the ability to deploy instances of the storybook app to an ECS Fargate cluster from feature branches. These can be created as part of the pull request workflow in travisCI and then torn down when the branch is merged. This should provide useful functionality with little additional overhead for the development team.

These are early versions of these scripts, which need to be updated to be executable with a single `npm` command and to better handle errors / async functionality. There also needs to be a script written that can tear down the containers once they are no longer needed.